requirements.txt
# VirtualRAG Dependencies

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0

# Vector Database & RAG
chromadb>=0.4.18
langchain>=0.1.0
langchain-community>=0.0.10

# Document Processing
pypdf>=3.17.0

# LLM Integration
requests>=2.31.0
ollama>=0.1.0 

# Utilities
pydantic>=2.5.0
